DashboardScene: Drop actions barrel file to satisfy lint rules
no-barrel-files forbids re-exporting imported bindings in public/app, so
remove actions/index.ts and keep the dashboardEditActions facade declared
in sidebar/shared.ts. getEditableElementFor consumers now import it from
actions/utils directly.
