--- public/app/features/dashboard-scene/actions/index.ts
@@ -1,12 +1,0 @@
-export { dashboardEditActions } from './dashboardEditActions';
-export { getEditableElementFor } from './utils/getEditableElementFor';
-export type {
-  AddElementActionHelperProps,
-  AddVariableActionHelperProps,
-  ChangeVariableTypeActionHelperProps,
-  EditActionProps,
-  MakeEditActionProps,
-  MoveElementActionHelperProps,
-  RemoveElementActionHelperProps,
-  RemoveVariableActionHelperProps,
-} from './utils/types';
