+++ public/app/features/dashboard-scene/sidebar/outline/DashboardOutlineNode.tsx
@@ -12,6 +12,7 @@ import {
   HIGHLIGHT_CATEGORY_PARAM_NAME,
 } from 'app/features/dashboard/components/PanelEditor/OptionsPaneCategory';
 
+import { getEditableElementFor } from '../../actions/utils/getEditableElementFor';
 import { DashboardDataLayerSet } from '../../scene/DashboardDataLayerSet';
 import { DashboardScene } from '../../scene/DashboardScene';
 import { RowItem } from '../../scene/layout-rows/RowItem';
@@ -23,7 +24,6 @@ import { SectionFiltersSet } from '../../settings/variables/SectionFiltersSet';
 import { isRepeatCloneOrChildOf } from '../../utils/clone';
 import { DashboardInteractions } from '../../utils/interactions';
 import { getDashboardSceneFor } from '../../utils/utils';
-import { getEditableElementFor } from '../shared';
 import { SidebarCategoryType, type DashboardSidebarLike } from '../types';
 import { useOutlineRename } from '../useOutlineRename';
