+++ public/app/features/dashboard-scene/sidebar/shared.ts
@@ -2,23 +2,41 @@ import { useSessionStorage } from 'react-use';
 
 import { type ElementSelectionContextItem } from '@grafana/ui';
 
+import { changeDescription } from '../actions/dashboard/changeDescription';
+import { changeTitle } from '../actions/dashboard/changeTitle';
+import { addElement } from '../actions/element/addElement';
+import { moveElement } from '../actions/element/moveElement';
+import { removeElement } from '../actions/element/removeElement';
+import { edit } from '../actions/utils/edit';
 import { getEditableElementFor } from '../actions/utils/getEditableElementFor';
+import { addVariable } from '../actions/variable/addVariable';
+import { changeVariableDescription } from '../actions/variable/changeVariableDescription';
+import { changeVariableHideValue } from '../actions/variable/changeVariableHideValue';
+import { changeVariableLabel } from '../actions/variable/changeVariableLabel';
+import { changeVariableName } from '../actions/variable/changeVariableName';
+import { changeVariableType } from '../actions/variable/changeVariableType';
+import { removeVariable } from '../actions/variable/removeVariable';
 import { type BulkActionElement, isBulkActionElement } from '../scene/types/BulkActionElement';
 import { type EditableDashboardElement } from '../scene/types/EditableDashboardElement';
 
 import { type DashboardSidebar } from './DashboardSidebar';
 import { MultiSelectedObjectsEditableElement } from './MultiSelectedObjectsEditableElement';
 
-export { dashboardEditActions } from '../actions';
-export { getEditableElementFor } from '../actions/utils/getEditableElementFor';
-export type {
-  AddElementActionHelperProps,
-  AddVariableActionHelperProps,
-  ChangeVariableTypeActionHelperProps,
-  MoveElementActionHelperProps,
-  RemoveElementActionHelperProps,
-  RemoveVariableActionHelperProps,
-} from '../actions/utils/types';
+export const dashboardEditActions = {
+  edit,
+  addElement,
+  removeElement,
+  moveElement,
+  changeTitle,
+  changeDescription,
+  addVariable,
+  removeVariable,
+  changeVariableType,
+  changeVariableName,
+  changeVariableLabel,
+  changeVariableDescription,
+  changeVariableHideValue,
+};
 
 export const SIDEBAR_COLLAPSED_KEY = 'grafana.dashboards.sidebar.isCollapsed';
